# backend/bdd/doc/openapi/bddOpenApi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Generated API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:5001"
  } ],
  "paths" : {
...
    "/api/database/restaurants/create" : {
      "post" : {
        "summary" : "createRestaurant",
        "description" : "Generated from method createRestaurant",
        "parameters" : [ {
          "name" : "RestaurantDTO",
          "in" : "body",
          "schema" : {
            "$ref" : "#/components/schemas/RestaurantDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/database/restaurants/{restaurantId}/menuItem/{menuItemId}" : {
      "get" : {
        "summary" : "getMenuItem",
        "description" : "Generated from method getMenuItem",
        "parameters" : [ {
          "name" : "restaurantId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "object",
            "properties" : {
              "serialVersionUID" : {
                "type" : "integer"
              },
              "mostSigBits" : {
                "type" : "integer"
              },
              "leastSigBits" : {
                "type" : "integer"
              },
              "jla" : {
                "type" : "object"
              },
              "NIBBLES" : {
                "type" : "object"
              },
              "$assertionsDisabled" : {
                "type" : "boolean"
              }
            }
          }
        }, {
          "name" : "menuItemId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "object",
            "properties" : {
              "serialVersionUID" : {
                "type" : "integer"
              },
              "mostSigBits" : {
                "type" : "integer"
              },
              "leastSigBits" : {
                "type" : "integer"
              },
              "jla" : {
                "type" : "object"
              },
              "NIBBLES" : {
                "type" : "object"
              },
              "$assertionsDisabled" : {
                "type" : "boolean"
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/database/restaurants/update" : {
      "put" : {
        "summary" : "updateRestaurant",
        "description" : "Generated from method updateRestaurant",
        "parameters" : [ {
          "name" : "RestaurantDTO",
          "in" : "body",
          "schema" : {
            "$ref" : "#/components/schemas/RestaurantDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/database/restaurants/{id}" : {
      "get" : {
        "summary" : "getRestaurantById",
        "description" : "Generated from method getRestaurantById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "object",
            "properties" : {
              "serialVersionUID" : {
                "type" : "integer"
              },
              "mostSigBits" : {
                "type" : "integer"
              },
              "leastSigBits" : {
                "type" : "integer"
              },
              "jla" : {
                "type" : "object"
              },
              "NIBBLES" : {
                "type" : "object"
              },
              "$assertionsDisabled" : {
                "type" : "boolean"
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "deleteRestaurant",
        "description" : "Generated from method deleteRestaurant",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "object",
            "properties" : {
              "serialVersionUID" : {
                "type" : "integer"
              },
              "mostSigBits" : {
                "type" : "integer"
              },
              "leastSigBits" : {
                "type" : "integer"
              },
              "jla" : {
                "type" : "object"
              },
              "NIBBLES" : {
                "type" : "object"
              },
              "$assertionsDisabled" : {
                "type" : "boolean"
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/database/managers/{id}/restaurant/{restaurantId}" : {
      "get" : {
        "summary" : "getManagerRestaurant",
        "description" : "Generated from method getManagerRestaurant",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "object",
            "properties" : {
              "serialVersionUID" : {
                "type" : "integer"
              },
              "mostSigBits" : {
                "type" : "integer"
              },
              "leastSigBits" : {
                "type" : "integer"
              },
              "jla" : {
                "type" : "object"
              },
              "NIBBLES" : {
                "type" : "object"
              },
              "$assertionsDisabled" : {
                "type" : "boolean"
              }
            }
          }
        }, {
          "name" : "restaurantId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "object",
            "properties" : {
              "serialVersionUID" : {
                "type" : "integer"
              },
              "mostSigBits" : {
                "type" : "integer"
              },
              "leastSigBits" : {
                "type" : "integer"
              },
              "jla" : {
                "type" : "object"
              },
              "NIBBLES" : {
                "type" : "object"
              },
              "$assertionsDisabled" : {
                "type" : "boolean"
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "HttpResponse" : {
        "type" : "object",
        "properties" : {
          "statusCode" : {
            "type" : "integer"
          },
          "body" : {
            "type" : "string"
          }
        }
      },
...
      "RestaurantDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "restaurantName" : {
            "type" : "string"
          },
          "menu" : {
            "type" : "object"
          },
          "slots" : {
            "type" : "array"
          },
          "openingTime" : {
            "type" : "string"
          },
          "closingTime" : {
            "type" : "string"
          },
          "orOrderPriceStrategy" : {
            "type" : "object"
          },
          "pendingOrders" : {
            "type" : "object"
          }
        }
      },
...
    }
  }
}